feat(0.3.1): Added version support for generating michelson
