+++ versioned/version0_0_6.ts
@@ -8,7 +8,6 @@ import {
 } from "@taquito/taquito";
 import { BigNumber } from "bignumber.js";
 import { DEFAULT_TIMEOUT } from "../context/config";
-import { makeFa2Michelson } from "../context/fa2";
 import { content, contractStorage as storage } from "../types/Proposal0_0_6";
 import { contractStorage } from "../types/app";
 import { proposal, proposalContent, status } from "../types/display";
@@ -54,7 +53,8 @@ class Version0_0_6 extends Versioned {
               const parser = new Parser();
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa2Michelson(
+                Versioned.generateFA2Michelson(
+                  this.version,
                   x.values.map(value => ({
                     walletAddress: cc.address,
                     targetAddress: value.targetAddress,
