+++ versioned/version0_3_0.ts
@@ -16,9 +16,6 @@ import { BigNumber } from "bignumber.js";
 import { fa1_2Token } from "../components/FA1_2";
 import { fa2Token } from "../components/FA2Transfer";
 import { DEFAULT_TIMEOUT } from "../context/config";
-import { makeFa1_2ApproveMichelson } from "../context/fa1_2";
-import { makeFa1_2TransferMichelson } from "../context/fa1_2";
-import { makeFa2Michelson } from "../context/fa2";
 import {
   content,
   proposal as p1,
@@ -214,7 +211,8 @@ class Version0_3_0 extends Versioned {
               const parser = new Parser();
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa2Michelson(
+                Versioned.generateFA2Michelson(
+                  this.version,
                   x.values.map(value => {
                     const token = value.token as unknown as fa2Token;
 
@@ -255,7 +253,7 @@ class Version0_3_0 extends Versioned {
               const token = x.values.token as unknown as fa1_2Token;
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa1_2ApproveMichelson({
+                Versioned.generateFA1_2ApproveMichelson(this.version, {
                   spenderAddress: x.values.spenderAddress,
                   amount: BigNumber(x.values.amount)
                     .multipliedBy(
@@ -289,7 +287,7 @@ class Version0_3_0 extends Versioned {
               const token = x.values.token as unknown as fa1_2Token;
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa1_2TransferMichelson({
+                Versioned.generateFA1_2TransferMichelson(this.version, {
                   walletAddress: cc.address,
                   amount: BigNumber(x.values.amount)
                     .multipliedBy(
