+++ versioned/version0_3_1.ts
@@ -10,9 +10,6 @@ import { BigNumber } from "bignumber.js";
 import { fa1_2Token } from "../components/FA1_2";
 import { fa2Token } from "../components/FA2Transfer";
 import { DEFAULT_TIMEOUT } from "../context/config";
-import { makeFa1_2ApproveMichelson } from "../context/fa1_2";
-import { makeFa1_2TransferMichelson } from "../context/fa1_2";
-import { makeFa2Michelson } from "../context/fa2";
 import {
   content,
   proposal as p1,
@@ -80,7 +77,8 @@ class Version0_3_0 extends Versioned {
               const parser = new Parser();
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa2Michelson(
+                Versioned.generateFA2Michelson(
+                  this.version,
                   x.values.map(value => {
                     const token = value.token as unknown as fa2Token;
 
@@ -121,7 +119,7 @@ class Version0_3_0 extends Versioned {
               const token = x.values.token as unknown as fa1_2Token;
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa1_2ApproveMichelson({
+                Versioned.generateFA1_2ApproveMichelson(this.version, {
                   spenderAddress: x.values.spenderAddress,
                   amount: BigNumber(x.values.amount)
                     .multipliedBy(
@@ -155,7 +153,7 @@ class Version0_3_0 extends Versioned {
               const token = x.values.token as unknown as fa1_2Token;
 
               const michelsonCode = parser.parseMichelineExpression(
-                makeFa1_2TransferMichelson({
+                Versioned.generateFA1_2TransferMichelson(this.version, {
                   walletAddress: cc.address,
                   amount: BigNumber(x.values.amount)
                     .multipliedBy(
@@ -311,8 +309,9 @@ class Version0_3_0 extends Versioned {
                   meta: content.execute_lambda.metadata
                     ? bytes2Char(content.execute_lambda.metadata)
                     : "No meta supplied",
+
                   lambda: emitMicheline(
-                    content.execute_lambda.lambda.map(str => JSON.parse(str))
+                    JSON.parse(content.execute_lambda.lambda ?? "")
                   ),
                 },
                 null,
@@ -329,11 +328,7 @@ class Version0_3_0 extends Versioned {
                 2
               ),
           content: content.execute_lambda.lambda
-            ? emitMicheline(
-                (content.execute_lambda.lambda ?? [""]).map(str =>
-                  JSON.parse(str)
-                )
-              )
+            ? emitMicheline(JSON.parse(content.execute_lambda.lambda))
             : "",
         },
       };
